pyproject.toml
[tool.poetry]
name = "ghostpost-tlwilliams895"
version = "0.1.0"
description = "GhostPost Project"
authors = ["tlwilliams895 <tlwilliams895@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "Using version ^3.1.5 for Django"
django-bootstrap4 = "^3.0.1"
black = "^21.5b1"
GitPython = "^3.1.16"
python-decouple = "^3.4"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
ipdb = "Using version ^0.13.4 for ipdb"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
